feat: Update user profile to allow profile picture and name updates
This commit updates the user profile functionality to allow users to update their profile picture and name. It adds the `publicPath` property to the webpack configuration to set the root path for static assets. In the `index.html` file, the paths for the favicon and bundle.js are updated to include the root path. In the `urls.py` file, the `get_user_profile` API endpoint is modified to use the `path` function instead of the `url` function for consistency. In the `serializers.py` file, the `ImageModelSerializer` class is commented out. In the `profile.js` file, the `bootstrap` import is removed and the `showToast` import is moved below the `ProfileInfo` import. The `Profile` class is updated to use the `post` method instead of the `put` method for updating the user profile. The `UserUpdateView` class is updated to use the `post` method instead of the `put` method for handling the update of the user profile. The `first_name` and `last_name` variables are updated to use the French names `nom` and `prenom`. The `print` statement is removed from the `create` method of the `ImageModelSerializer` class. The commit also includes various formatting and code cleanup changes.